requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
WTForms==3.0.1
Flask-WTF==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
SQLAlchemy==2.0.19
python-dotenv==1.2.1
pymysql==1.1.0
